gcloud_webhook/requirements.txt
python-telegram-bot==22.0
groq==0.23.1
uvicorn==0.23.2
starlette==0.37.2
httpx==0.28.1
google-cloud-storage==2.14.0
google-cloud-secret-manager==2.20.0